Python/azure_openai/Langchain_Qdrant_Web_API/pyproject.toml
[tool.poetry]
name = "hw_app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = ">2"
# pydantic = "<2"
langchain-openai = "^0.2.8"
langchain-community = "^0.3.7"
qdrant-client = "^1.12.1"
langchain-qdrant = "^0.2.0"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
